# openapi/product.json
{
  "components": {
    "schemas": {
...
      "NormalizedWorkflow-Input": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "maxLength": 1024,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Description",
            "type": "string"
          },
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9]*(?:[-_][a-z0-9]+)*$",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "const": "workflow",
            "title": "Kind",
            "type": "string"
          },
          "lead": {
            "$ref": "#/components/schemas/NormalizedMember-Input"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 8192,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "kind",
          "id",
          "description",
          "lead"
        ],
        "title": "NormalizedWorkflow",
        "type": "object"
      },
...
      "OperationFailure": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProductFailureCode"
          },
          "field_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Path"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "suggested_next_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Next Step"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "code",
          "summary",
          "retryable"
        ],
        "title": "OperationFailure",
        "type": "object"
      },
...
      "WorkflowDraftReadback": {
        "additionalProperties": false,
        "properties": {
          "base_revision_no": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Revision No"
          },
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "etag": {
            "title": "Etag",
            "type": "string"
          },
          "workflow": {
            "$ref": "#/components/schemas/NormalizedWorkflow-Output"
          },
          "workflow_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9]*(?:[-_][a-z0-9]+)*$",
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "draft_id",
          "workflow_id",
          "etag",
          "workflow"
        ],
        "title": "WorkflowDraftReadback",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Banksia Product API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/workflow-drafts": {
      "post": {
        "operationId": "post_workflow_draft_workflow_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NormalizedWorkflow-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDraftReadback"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Post Workflow Draft",
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflow-drafts/{draft_id}": {
      "delete": {
        "operationId": "delete_workflow_draft_workflow_drafts__draft_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDraftDiscardResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStaleDraftResponse"
                }
              }
            },
            "description": "Precondition Failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowPreconditionRequiredResponse"
                }
              }
            },
            "description": "Precondition Required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Delete Workflow Draft",
        "tags": [
          "workflows"
        ]
      },
      "get": {
        "operationId": "get_workflow_draft_workflow_drafts__draft_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDraftReadback"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Workflow Draft",
        "tags": [
          "workflows"
        ]
      },
      "patch": {
        "operationId": "patch_workflow_draft_workflow_drafts__draft_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "add_member": "#/components/schemas/AddMemberOperation",
                    "remove_member": "#/components/schemas/RemoveMemberOperation",
                    "update_member": "#/components/schemas/UpdateMemberOperation",
                    "update_workflow": "#/components/schemas/UpdateWorkflowOperation"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/UpdateWorkflowOperation"
                  },
                  {
                    "$ref": "#/components/schemas/AddMemberOperation"
                  },
                  {
                    "$ref": "#/components/schemas/UpdateMemberOperation"
                  },
                  {
                    "$ref": "#/components/schemas/RemoveMemberOperation"
                  }
                ],
                "title": "Operation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDraftMutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStaleDraftResponse"
                }
              }
            },
            "description": "Precondition Failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowPreconditionRequiredResponse"
                }
              }
            },
            "description": "Precondition Required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Patch Workflow Draft",
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflow-drafts/{draft_id}/publish": {
      "post": {
        "operationId": "post_workflow_draft_publish_workflow_drafts__draft_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowPublishedReadback"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStaleDraftResponse"
                }
              }
            },
            "description": "Precondition Failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowPreconditionRequiredResponse"
                }
              }
            },
            "description": "Precondition Required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Post Workflow Draft Publish",
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflow-drafts/{draft_id}/undo": {
      "post": {
        "operationId": "post_workflow_draft_undo_workflow_drafts__draft_id__undo_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowUndoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDraftReadback"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStaleDraftResponse"
                }
              }
            },
            "description": "Precondition Failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowPreconditionRequiredResponse"
                }
              }
            },
            "description": "Precondition Required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Post Workflow Draft Undo",
        "tags": [
          "workflows"
        ]
      }
    },
...
  }
}